+++ devicemodel/hw/platform/ioc.c
@@ -55,7 +55,6 @@
 #include <unistd.h>
 #include <errno.h>
 #include <fcntl.h>
-#include <pty.h>
 #include <string.h>
 #include <stdbool.h>
 #include <types.h>
@@ -63,6 +62,8 @@
 
 #include <sys/stat.h>
 #include <sys/types.h>
+
+#include "pty_vuart.h"
 
 #include "dm.h"
 #include "ioc.h"
@@ -710,83 +711,6 @@ ioc_open_native_ch(const char *dev_name)
 }
 
 /*
- * Check and create the directory.
- * To avoid symlink failure if the directory does not exist.
- */
-static int
-check_dir(const char *file)
-{
-	char *tmp, *dir;
-
-	tmp = strdup(file);
-	if (!tmp) {
-		DPRINTF("ioc falied to dup file, error:%s\r\n",
-				strerror(errno));
-		return -1;
-	}
-
-	dir = dirname(tmp);
-	if (access(dir, F_OK) && mkdir(dir, 0666)) {
-		DPRINTF("ioc falied to create dir:%s, erorr:%s\r\n", dir,
-				strerror(errno));
-		free(tmp);
-		return -1;
-	}
-	free(tmp);
-	return 0;
-}
-
-/*
- * Open PTY master device for IOC mediator and the PTY slave device for virtual
- * UART. The pair(master/slave) can work as a communication channel between
- * IOC mediator and virtual UART.
- */
-static int
-ioc_open_virtual_uart(const char *dev_name)
-{
-	int fd;
-	char *slave_name;
-	struct termios attr;
-
-	fd = open("/dev/ptmx", O_RDWR | O_NOCTTY | O_NONBLOCK);
-	if (fd < 0)
-		goto open_err;
-	if (grantpt(fd) < 0)
-		goto pty_err;
-	if (unlockpt(fd) < 0)
-		goto pty_err;
-	slave_name = ptsname(fd);
-	if (!slave_name)
-		goto pty_err;
-	if ((unlink(dev_name) < 0) && errno != ENOENT)
-		goto pty_err;
-	/*
-	 * The check_dir restriction is that only create one directory
-	 * not support multi-level directroy.
-	 */
-	if (check_dir(dev_name) < 0)
-		goto pty_err;
-	if (symlink(slave_name, dev_name) < 0)
-		goto pty_err;
-	if (chmod(dev_name, 0660) < 0)
-		goto attr_err;
-	if (tcgetattr(fd, &attr) < 0)
-		goto attr_err;
-	cfmakeraw(&attr);
-	attr.c_cflag |= CLOCAL;
-	if (tcsetattr(fd, TCSANOW, &attr) < 0)
-		goto attr_err;
-	return fd;
-
-attr_err:
-	unlink(dev_name);
-pty_err:
-	close(fd);
-open_err:
-	return -1;
-}
-
-/*
  * Open native CBC cdevs and virtual UART.
  */
 static int
@@ -807,7 +731,7 @@ ioc_ch_init(struct ioc_dev *ioc)
 			fd = ioc_open_native_ch(chl->name);
 			break;
 		case IOC_VIRTUAL_UART:
-			fd = ioc_open_virtual_uart(virtual_uart_path);
+			fd = pty_open_virtual_uart(virtual_uart_path);
 			break;
 		case IOC_LOCAL_EVENT:
 			if (!pipe(pipe_fds)) {
@@ -823,7 +747,7 @@ ioc_ch_init(struct ioc_dev *ioc)
 		 */
 		case IOC_NATIVE_DUMMY0:
 			if (ioc_debug_enable) {
-				fd = ioc_open_virtual_uart(chl->name);
+				fd = pty_open_virtual_uart(chl->name);
 				dummy0_sfd = open(chl->name, O_RDWR | O_NOCTTY |
 						O_NONBLOCK);
 			} else
@@ -831,7 +755,7 @@ ioc_ch_init(struct ioc_dev *ioc)
 			break;
 		case IOC_NATIVE_DUMMY1:
 			if (ioc_debug_enable) {
-				fd = ioc_open_virtual_uart(chl->name);
+				fd = pty_open_virtual_uart(chl->name);
 				dummy1_sfd = open(chl->name, O_RDWR | O_NOCTTY |
 						O_NONBLOCK);
 			} else
@@ -839,7 +763,7 @@ ioc_ch_init(struct ioc_dev *ioc)
 			break;
 		case IOC_NATIVE_DUMMY2:
 			if (ioc_debug_enable) {
-				fd = ioc_open_virtual_uart(chl->name);
+				fd = pty_open_virtual_uart(chl->name);
 				dummy2_sfd = open(chl->name, O_RDWR | O_NOCTTY |
 						O_NONBLOCK);
 			} else
