+++ devicemodel/include/pty_vuart.h
@@ -1,0 +1,16 @@
+/*
+ * Project Acrn
+ * Acrn-dm-pty
+ *
+ * Copyright (C) 2019 Intel Corporation. All rights reserved.
+ *
+ * SPDX-License-Identifier: BSD-3-Clause
+ *
+ */
+
+#ifndef __PTY_H__
+#define __PTY_H__
+
+int pty_open_virtual_uart(const char *dev_name);
+
+#endif
